requirements-test.txt
-r requirements.txt
colorlog~=6.8
flake8~=7.0
flake8-docstrings~=1.7
fnv-hash-fast~=0.4
mypy~=1.9
psutil-home-assistant==0.0.1
pylint~=3.1
pylint-strict-informational==0.1
pytest>=7.2
pytest-cov>=3.0
pytest-homeassistant-custom-component>=0.12
pytest-asyncio>=0.20
tzdata
ruff~=0.3
